% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiproduct.R
\name{AttributeRisk}
\alias{AttributeRisk}
\title{AttributeRisk}
\usage{
AttributeRisk(
  modelFormulas,
  origdata,
  syndata,
  posteriorMCMCs,
  syntype,
  H = 50,
  G = 11,
  percentBounds = c(0.1, 0.1),
  additiveBounds = NULL,
  bounds = NULL,
  guesses = NULL,
  simplePrior = NULL,
  categorical = NULL
)
}
\value{
a list of lists (one for each record), where each inner list contains the following:
\itemize{
\item The full probability matrix
\item Marginal probabilities
\item The true value probability
\item The ranking of the guess with the highest probability
\item The ranking of the true value
}
}
\description{
AttributeRisk
}
