% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiproduct.R
\name{AttributeRiskForRecordI}
\alias{AttributeRiskForRecordI}
\title{AttributeRiskForRecordI}
\usage{
AttributeRiskForRecordI(
  modelFormulas,
  i,
  origdata,
  syndata,
  posteriorMCMCs,
  syntype,
  H = 50,
  G = 11,
  percentBounds = c(0.1, 0.1),
  additiveBounds = NULL,
  bounds = NULL,
  guesses = NULL,
  simplePrior = NULL,
  categorical = NULL
)
}
\description{
AttributeRiskForRecordI
}
