% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiproduct.R
\name{AttributeRiskForRecordI_multiSynthproduct}
\alias{AttributeRiskForRecordI_multiSynthproduct}
\title{AttributeRiskForRecordI_multiSynthproduct}
\usage{
AttributeRiskForRecordI_multiSynthproduct(
  modelFormulas,
  i,
  origdata,
  syndata,
  posteriorMCMCs,
  syntype,
  H = 50,
  G = 11,
  percentBounds = c(0.1, 0.1),
  additiveBounds = NULL,
  bounds = NULL,
  guesses = NULL,
  simplePrior = NULL
)
}
\description{
AttributeRiskForRecordI_multiSynthproduct
}
