% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiproduct.R
\name{randomGuessPlot}
\alias{randomGuessPlot}
\title{randomGuessPlot}
\usage{
randomGuessPlot(risks, custom_palette = NULL)
}
\arguments{
\item{risks}{List containing the true value probabilities (output from \code{\link[AttributeRiskCalculation]{AttributeRisk}})}

\item{custom_palette}{Vector of color values to use for the plot. Will use the first value for the density and the last for the vertical line.}
}
\value{
A ggplot plot object containing the desired plot.
}
\description{
Produces a graph of the probabilities of each guess with a line indicating 
the chance of randomly guessing the confidential value(s) from among the 
guesses.
}
